+++ CHANGELOG.md
@@ -30,6 +30,8 @@ All notable changes to this project will be documented in this file. It uses the
 *   Fixed foreign-table and user-mapping selection for pushed-down join and
     upper-relation scans ([#319]).
 *   Fixed subsecond precision lost inserting timestamps over HTTP ([#300]).
+*   Fixed loss of subsecond precision in `to_timestamp(float8)` by mapping it
+    to `toDateTime64()`. Thanks to Minh Vu for the PR ([#338]).
 
   [#296]: https://github.com/ClickHouse/pg_clickhouse/pull/296
     "ClickHouse/pg_clickhouse#296 Fix benchmark queries that crash/hang with binary driver"
@@ -39,6 +41,9 @@ All notable changes to this project will be documented in this file. It uses the
     "ClickHouse/pg_clickhouse#324 Push down compatible UNION queries and grouped set operations"
   [#300]: https://github.com/ClickHouse/pg_clickhouse/pull/300
     "ClickHouse/pg_clickhouse#300 fix(http): handle subsecond precision"
+
+  [#338]: https://github.com/ClickHouse/pg_clickhouse/pull/338
+    "ClickHouse/pg_clickhouse#338 Preserve fractional seconds in to_timestamp()"
 
 ## [v0.3.2] — 2026-06-16